refactor: create subcomponent CompactNotification
