+++ src/components/Notification/subcomponents/CompactNotification/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './CompactNotification';
